+++ src/Interface/AppPreferences/AppPreferences.h
@@ -90,7 +90,7 @@ private Q_SLOTS:
   
   // SET_TIMER_LABEL:
   // This function is called when the slider is updated and it shows the actual value of the slider
-  void set_timer_label( double );
+  void set_timer_label( int );
   
 private:
   boost::shared_ptr< AppPreferencesPrivate > private_;
